Use webpack for photography CSS and JS
