+++ app/views/layouts/application.html.slim
@@ -9,6 +9,7 @@ html lang='en'
     = auto_discovery_link_tag :atom, blog_posts_path(format: :atom)
     = stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track' => true
     = javascript_pack_tag 'application', 'data-turbolinks-track': 'reload'
+    = yield :head
     = csrf_meta_tags
   body class="#{@body_class}"
     .site
